Wrote in predictive current limitng in arm, untested
Predictive current limiting uses a model of a motor to see if it
will go over a set current limit, if it does, it will dial back
the voltage. This makes things like PID loops faster and
smoother. Rather than waiting for current to go over and turning
it down it ensures that it never goes over at all. This method
does require an encoder on though, and I have also kept the
previous limits, although they shouldn't ever be hit.